Move outreach test fixtures to root confest.py.
