+++ users/tests/test_admin.py
@@ -1,37 +1,9 @@
-import pytest
-from django.test.client import Client
-from django.contrib.auth.models import Group
-from django.core.management import call_command
-
 from .factories import UserFactory
 
 
 # Content that shows up on a user change page if the
 # logged-in user can set the superuser status of the user.
 SUPERUSER_SENTINEL = 'superuser'
-
-
-@pytest.fixture
-def initgroups(db):
-    call_command('initgroups')
-
-
-@pytest.fixture
-def outreach_user(initgroups):
-    user = UserFactory(
-        username='outreacher',
-        phone_number='1234567000',
-        is_staff=True)
-    group = Group.objects.get(name='Outreach Coordinators')
-    user.groups.add(group)
-    return user
-
-
-@pytest.fixture
-def outreach_client(outreach_user):
-    client = Client()
-    client.force_login(outreach_user)
-    return client
 
 
 def test_list_view_works(admin_client):
